requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.0
accelerate>=0.20.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pypdf>=3.0.0

# Text processing
nltk>=3.8
spacy>=3.6.0
tiktoken>=0.5.0

# Configuration
PyYAML>=6.0
pydantic>=2.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
colorlog>=6.7.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Optional: For advanced PDF processing
pdf2image>=1.16.0
pillow>=10.0.0

# Optional: For better performance
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4